Rework Endoflame
Now it consumes all nearby burnables to generate mana immediately but takes time to transfer.
It adopts a progressive formula where it generates more mana the less work it is given.
Transfer rate works the opposite, with the flower overclocking itself at the cost of efficiency.
The flower also loses mana based on the current capacity.
Capacity | Efficiency | Transfer Speed | Mana loss
<25% | 1.6x | 8/s | 2/s
<50% | 0.8x |16/s | 4/s
<75% | 0.4x | 32/s | 8/s
<100% | 0.2x | 64/s | 16/s

Overworking the flower leads to it getting burned out, a cooldown lasting the excess heat is dispersed (up to 64 second). This is shown both as particles and in the GUI.

TL;DR: Mana produced is exponentially higher in the early quarter but slow, Mana output is exponentially faster in the later quarters but inefficient